backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
Pillow==10.4.0
 django-cors-headers==4.4.0

django-admin-decorators==0.1
jdatetime==5.0.0

#test
pytest==8.3.2
pytest-django==4.8.0